requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.0
coverage==7.12.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.23.1
psycopg==3.2.13
psycopg-binary==3.2.13
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.3.3
pytest-cov==5.0.0
python-jose==3.3.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
passlib[argon2]
psycopg[binary]>=3.1
psycopg2-binary
